flightlog: add schema and related methods
Flight log schema has been added, this schema is supports all events
including takeoff, landing, and bundling. This commit adds a basic
schema validator using jsonchema and adds a unittest for the same.
Note that the schema validation is not included in the checks performed
by the apps in this commit.

A sample flightlog that conforms to the new schema is also provided.

Requirements.txt is updated to include jsonschema